[Grupo 12] - req 13

+++ back-end/src/database/models/product.js
@@ -1,7 +1,7 @@
 module.exports = (sequelize, DataTypes) => {
   const Product = sequelize.define('Product', {
     name: DataTypes.STRING,
-    price: DataTypes.FLOAT,
+    price: DataTypes.DECIMAL,
     urlImage: DataTypes.STRING,
   }, { tableName: 'products', timestamps: false });
